requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
ale-py==0.9.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argcomplete==3.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
AutoROM==0.6.1
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto==2.49.0
cachetools==5.3.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
cmake==3.25.0
comm==0.2.2
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
dotmap==1.3.30
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
gcs-oauth2-boto-plugin==3.2
gitdb==4.0.11
GitPython==3.1.43
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-reauth==0.1.1
gsutil==5.30
gym==0.25.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.20.4
httpx==0.27.0
hydra-core==1.3.2
idna==3.4
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lit==15.0.7
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
omegaconf==2.3.0
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyu2f==0.1.5
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
retry_decorator==1.1.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
rsa==4.7.2
Send2Trash==1.8.3
sentry-sdk==2.9.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==1.26.13
wandb==0.17.4
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
yarl==1.9.4
zipp==3.19.2
